use filtred images to test resnet model and add gan for these data
As we have fit files converted to png and filtred with acs tools, we are able to check its accuracy and if gan behaves different with new file format. unfortunately accuracy stays the same as before, so gan doesnt change its results - we get grey images as a output. probably we need to train gan model